refactor : Reorganize Component Dir Sturucture
