+++ apps-example/international-example/src/components/SemanticBox.tsx
