+++ apps-example/international-example/src/components/collection/CollectionList.tsx
@@ -1,11 +1,12 @@
-import { ToolkitProduct } from '@/@marulloc-shopify-nextapi/v24.01/services/@toolkit-types/toolkit-product';
+import { ToolkitCollection } from '@/@marulloc-shopify-nextapi/v24.01/services/@toolkit-types/toolkit-collection';
+import CollectionCard, { TCollectionCardProps } from './CollectionCard';
 import { classNames } from '@marulloc/components-library/utils';
-import Link, { LinkProps } from 'next/link';
-import ProductCard, { TPRoductCardProps } from '../product/ProductCard';
 
-type TProductListProps = {
-  products: ToolkitProduct[];
-  variant: TPRoductCardProps['variant'];
+import Link, { LinkProps } from 'next/link';
+
+export type TCollectionListProps = {
+  collections: ToolkitCollection[];
+  variant: TCollectionCardProps['variant'];
   subProps?: {
     ul?: Omit<React.ComponentPropsWithoutRef<'ul'>, 'children'>;
     li?: Omit<React.ComponentPropsWithoutRef<'li'>, 'children'>;
@@ -13,7 +14,7 @@ type TProductListProps = {
   };
 };
 
-const ProductList = ({ products, variant, subProps }: TProductListProps) => {
+const CollectionList = ({ collections, variant, subProps }: TCollectionListProps) => {
   const uniqueId = `id-${new Date().getTime()}-${Math.random().toString(36).substr(2, 9)}`;
 
   const ulProps = subProps?.ul;
@@ -22,17 +23,17 @@ const ProductList = ({ products, variant, subProps }: TProductListProps) => {
 
   const ulClassName =
     variant === 'big'
-      ? classNames('grid grid-cols-2 md:grid-cols-3 lg:grid-cols-4 gap-4 sm:gap-6 md:gap-8', ulProps?.className)
+      ? classNames('grid grid-cols-3 gap-4 sm:gap-6 md:gap-8', ulProps?.className)
       : classNames('space-y-4', ulProps?.className);
   const liClassName =
     variant === 'big' ? classNames('aspect-square', liProps?.className) : classNames('py-1', liProps?.className);
 
   return (
     <ul {...ulProps} className={ulClassName}>
-      {products.map((product) => (
-        <li key={`${uniqueId}-product-card-${product.handle}`} {...liProps} className={liClassName}>
-          <Link href={product.handleRoute} {...linkProps}>
-            <ProductCard variant={variant} product={product} />
+      {collections.map((collection, index) => (
+        <li key={`${uniqueId}-collection-card-${collection.handle}`} {...liProps} className={liClassName}>
+          <Link href={collection.handleRoute} {...linkProps}>
+            <CollectionCard variant={variant} collection={collection} index={index} />
           </Link>
         </li>
       ))}
@@ -40,4 +41,4 @@ const ProductList = ({ products, variant, subProps }: TProductListProps) => {
   );
 };
 
-export default ProductList;
+export default CollectionList;
