+++ apps-example/international-example/src/components/product/Price.tsx
