+++ apps-example/international-example/src/components/product/ProductList.tsx
@@ -1,12 +1,11 @@
-import { ToolkitCollection } from '@/@marulloc-shopify-nextapi/v24.01/services/@toolkit-types/toolkit-collection';
-import CollectionCard, { TCollectionCardProps } from '../collection/CollectionCard';
+import { ToolkitProduct } from '@/@marulloc-shopify-nextapi/v24.01/services/@toolkit-types/toolkit-product';
 import { classNames } from '@marulloc/components-library/utils';
-
 import Link, { LinkProps } from 'next/link';
+import ProductCard, { TPRoductCardProps } from './ProductCard';
 
-export type TCollectionListProps = {
-  collections: ToolkitCollection[];
-  variant: TCollectionCardProps['variant'];
+type TProductListProps = {
+  products: ToolkitProduct[];
+  variant: TPRoductCardProps['variant'];
   subProps?: {
     ul?: Omit<React.ComponentPropsWithoutRef<'ul'>, 'children'>;
     li?: Omit<React.ComponentPropsWithoutRef<'li'>, 'children'>;
@@ -14,7 +13,7 @@ export type TCollectionListProps = {
   };
 };
 
-const CollectionList = ({ collections, variant, subProps }: TCollectionListProps) => {
+const ProductList = ({ products, variant, subProps }: TProductListProps) => {
   const uniqueId = `id-${new Date().getTime()}-${Math.random().toString(36).substr(2, 9)}`;
 
   const ulProps = subProps?.ul;
@@ -23,17 +22,17 @@ const CollectionList = ({ collections, variant, subProps }: TCollectionListProps
 
   const ulClassName =
     variant === 'big'
-      ? classNames('grid grid-cols-3 gap-4 sm:gap-6 md:gap-8', ulProps?.className)
+      ? classNames('grid grid-cols-2 md:grid-cols-3 lg:grid-cols-4 gap-4 sm:gap-6 md:gap-8', ulProps?.className)
       : classNames('space-y-4', ulProps?.className);
   const liClassName =
     variant === 'big' ? classNames('aspect-square', liProps?.className) : classNames('py-1', liProps?.className);
 
   return (
     <ul {...ulProps} className={ulClassName}>
-      {collections.map((collection, index) => (
-        <li key={`${uniqueId}-collection-card-${collection.handle}`} {...liProps} className={liClassName}>
-          <Link href={collection.handleRoute} {...linkProps}>
-            <CollectionCard variant={variant} collection={collection} index={index} />
+      {products.map((product) => (
+        <li key={`${uniqueId}-product-card-${product.handle}`} {...liProps} className={liClassName}>
+          <Link href={product.handleRoute} {...linkProps}>
+            <ProductCard variant={variant} product={product} />
           </Link>
         </li>
       ))}
@@ -41,4 +40,4 @@ const CollectionList = ({ collections, variant, subProps }: TCollectionListProps
   );
 };
 
-export default CollectionList;
+export default ProductList;
