+++ apps-example/international-example/src/components/search/CollectionNav.tsx
@@ -1,5 +1,5 @@
 import { getCollections } from '@/@marulloc-shopify-nextapi/v24.01/services/collection/service';
-import CollectionList from '@/components/_draft/CollectionList';
+import CollectionList from '@/components/collection/CollectionList';
 import Skeleton from '@/components/loading/Skeleton';
 import { delay } from '@/utils/asyncUtils';
 import { Suspense } from 'react';
