+++ apps-example/international-example/src/components/shop/MenuDrawer/triggers/MenuIconTrigger.tsx
